Moved argv code to main() function
